% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities_cache.R
\name{cache.check}
\alias{cache.check}
\title{Check that the cache directory location has been set}
\usage{
cache.check()
}
\description{
Checks that the cache directory location has been set.  Also checks that the directory contains all necessary subdirectories
}

