% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utilities_cache.R
\name{cache.load.data}
\alias{cache.load.data}
\title{Downloads and/or loads data from cache directory}
\usage{
cache.load.data(file, dir,
  server.url = "http://felek.cns.umass.edu:9283/data", cache.only = F,
  col.names = NULL, message = "default", quiet = F, object = NULL,
  is.rdata = NULL)
}
\description{
Load data from cache directory.  If it does not exist, download it first and then load it.  Reassign the object within the parent environment
}
