% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utilities_cache.R
\name{cache.load.dataOLD}
\alias{cache.load.dataOLD}
\title{Downloads and/or loads data from cache directory}
\usage{
cache.load.dataOLD(file, dir, cache.only = F, is.rdata = T,
  col.names = NULL, server.url = "http://felek.cns.umass.edu:9283",
  message = "default", quiet = F, object = NULL)
}
\description{
Load data from cache directory.  If it does not exist, download it first and then load it.  Reassign the object within the parent environment
}
