% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities_cache.R
\name{cache.set}
\alias{cache.set}
\title{Sets the cache directory location}
\usage{
cache.set(cache.dir)
}
\arguments{
\item{cache.dir}{\code{character}  file path of cache directory location}
}
\description{
Sets the cache directory location.  Also checks that the directory conrtains all necessary subdirectories
}

