% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities_cache.R
\name{cache.setup}
\alias{cache.setup}
\title{Sets up directory structure for cache and temporary files}
\usage{
cache.setup(cache.dir, quiet = T)
}
\arguments{
\item{cache.dir}{\code{character} file path of cache directory location}

\item{quiet}{\code{should messages be suppressed}}
}
\description{
Creates directories with locations for cached catchment polygons, weather polygons, and weather data; log files; and temporary local storage
}

