% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/new cache functions.R
\name{cache_check}
\alias{cache_check}
\title{Check that the cache directory location has been set}
\usage{
cache_check(cache_dir = NULL, automatic_setup = F)
}
\description{
Checks that the cache directory location has been set.  Also checks that the directory contains all necessary subdirectories
}
