% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/new cache functions.R
\name{cache_download}
\alias{cache_download}
\title{Downloads and/or loads data from cache directory}
\usage{
cache_download(file, dir = "", server_url, quiet = F)
}
\description{
Load data from cache directory.  If it does not exist, download it first and then load it.  Reassign the object within the parent environment
}
