% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/new cache functions.R
\name{cache_setup}
\alias{cache_setup}
\title{Sets up directory structure for cache and temporary files}
\usage{
cache_setup(cache_dir, quiet = T)
}
\arguments{
\item{quiet}{\code{should messages be suppressed}}

\item{cache.dir}{\code{character} file path of cache directory location}
}
\description{
Creates directories with locations for cached catchment polygons, weather polygons, and weather data; log files; and temporary local storage
}
