% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{capitalize}
\alias{capitalize}
\title{Simple capitalization function}
\usage{
capitalize(string)
}
\arguments{
\item{string}{\code{character} string to capitalize}
}
\value{
character
}
\description{
Simple function to capitalize first letter or string, for prettying up plot titles, legends, axes, etc.
}

