% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clifs.periods.summary}
\alias{clifs.periods.summary}
\title{prints summary of CliFSSS model performance, prints detailed summary of  model for each period}
\usage{
clifs.periods.summary(ml, periods = NULL, generic = T, pause = F)
}
\arguments{
\item{ml}{\code{list} model list, of  \code{lm} or \code{merMod} class models}

\item{periods}{\code{array} of characters}

\item{generic}{\code{boolean}}

\item{pause}{\code{boolean} defaults to T, should the output be paused between each period}
}
\description{
prints summary of CliFSSS model performance, prints detailed summary of  model for each period
}

