% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather_deprecated.R
\name{create.cols.livneh}
\alias{create.cols.livneh}
\title{col names for livneh weather input data}
\usage{
create.cols.livneh()
}
\description{
col names for livneh weather data.  original files do not have column headers
}

