% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather_deprecated.R
\name{create.cols.mauer}
\alias{create.cols.mauer}
\title{col names for mauer weather input data}
\usage{
create.cols.mauer()
}
\description{
col names for mauer weather data.  original inputs do not have column headers
}

