% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_flow.R
\name{create.q.matrices}
\alias{create.q.matrices}
\title{create the monster list of 3-d matrices for flow}
\usage{
create.q.matrices(gages.spatial, periods = c("daily", "monthly", "seasonal",
  "annual"), template.date, template.period, cols.flow)
}
\description{
used to store aggregated flow data
}

