% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather_deprecated.R
\name{create.template.date.livnehA}
\alias{create.template.date.livnehA}
\title{Create template of dates in mauer weather data}
\usage{
create.template.date.livnehA()
}
\description{
livneh A dates 1915-2011
}

