% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather_deprecated.R
\name{create.template.date.livnehB}
\alias{create.template.date.livnehB}
\title{Create template of dates in mauer weather data}
\usage{
create.template.date.livnehB()
}
\description{
livneh B dates 1950-2013
}

