% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create.template.date}
\alias{create.template.date}
\title{Create template of dates}
\usage{
create.template.date(start.year = 1949, end.year = 2010)
}
\description{
Create template of dates to mark each period for both flow and weather data
}

