% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create.template.periods}
\alias{create.template.periods}
\title{define periods}
\usage{
create.template.periods(min.records.monthly = 25, min.records.seasonal = 80,
  min.records.annual = 345)
}
\description{
if we want to add different period (i.e. bkt bioperiod), can do it here
}

