% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather.R
\name{create.w.matrices}
\alias{create.w.matrices}
\title{Create monster list of 3d weather matrices}
\usage{
create.w.matrices(weather.filenames, template.date, template.period,
  cols.weather)
}
\description{
used to store aggregated weather metrics
}

