% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/validation.R
\name{cross.val}
\alias{cross.val}
\title{cross validate}
\usage{
cross.val(d.records, model.list, dep.var = "flow.mean",
  period.type = "month", iter = 10, valid.fraction = 0.1,
  valid.type = "mixed", seed = NULL, correction.iter = 100)
}
\description{
x
}
