% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data.merge}
\alias{data.merge}
\title{data merge
#}
\usage{
data.merge(gages.spatial, q.matrices = NULL, w.matrices = NULL,
  periods = c("seasonal"), select.gages = NULL, id.col = "site_no",
  filter.by = "flow.mean", template.period = NULL)
}
\description{
#
}

