% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utilities_cache.R
\name{dir.conditional.create}
\alias{dir.conditional.create}
\title{internal function to create directory if it doesn't already exist}
\usage{
dir.conditional.create(new.dir, quiet = T)
}
\arguments{
\item{new.dir}{\code{character}  character of new directory names}

\item{quiet}{\code{boolean}  whether to suppress messages}
}
