% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/new cache functions.R
\name{dir_create_conditional}
\alias{dir_create_conditional}
\title{internal function to create directory if it doesn't already exist}
\usage{
dir_create_conditional(new_dir, quiet = T)
}
\arguments{
\item{quiet}{\code{boolean}  whether to suppress messages}

\item{new.dir}{\code{character}  character of new directory names}
}
