% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_retrieve.R
\name{flow.pre.agg.function}
\alias{flow.pre.agg.function}
\title{calculations to perform on daily records before aggregation}
\usage{
flow.pre.agg.function(x.all)
}
\description{
by default, calculates rolling mean, and returns
}

