% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_retrieve.R
\name{flow.retrieve}
\alias{flow.retrieve}
\title{import flow}
\usage{
flow.retrieve(gages.spatial, periods = c("monthly", "annual"),
  flow.pre.agg.function = (conteStreamflow::flow.pre.agg.function),
  flow.agg.function = (conteStreamflow::flow.agg.function),
  start.year = 1890, end.year = 2015, min.records.monthly = 25,
  min.records.seasonal = 80, min.records.annual = 345, return.spatial = F)
}
\description{
import flow from nwis web service, and aggregate to various metrics and by various periods
}

