% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gage_retrieve.R
\name{gage.export}
\alias{gage.export}
\title{Export shapefile of gage information}
\usage{
gage.export(gages.spatial, shapefile.dir, filename = "gages_conteStreamflow",
  overwrite = T)
}
\description{
Export shapefile of gage information
}

