% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/gage_retrieve.R
\name{gage.retrieve}
\alias{gage.retrieve}
\title{Get gage location and information from NWIS}
\usage{
gage.retrieve(buffer.file = NULL, sites = NULL, states = NULL,
  max.da.sqkm = NULL, min.da.sqkm = NULL,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs",
  keep.other.coords = F, keep.sv.data = F)
}
\arguments{
\item{buffer.file}{\code{character} file name and location of a shapefile (polygon) that outlines area to find gages within}

\item{sites}{\code{character vector} of NWIS gage ID's ("gage_no")}

\item{states}{\code{character vector} of state abbreviations to search within.  This will only be used if buffer.file is not provided.}

\item{max.da.sqkm}{\code{numeric} filter gages by min and max drainage area}
}
\value{
\code{SpatialPointsDataFrame} of gages within the buffer, with gage info from NWIS
}
\description{
Get gage information, including site_no and lat/long coordinates from NWIS, by specifying either 
            1) a shapefile with a single polygon outlining an area of interest in the eastern U.S., or 
            2) a list of states to search within, or
            3) a vector of one or more NWIS gage IDs (as characters)
}
\keyword{gage}
\keyword{nwis,}
