% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/char_retrieve.R
\name{impound.retrieve}
\alias{impound.retrieve}
\title{impound retrieve}
\usage{
impound.retrieve(gages.spatial,
  server.url = "http://felek.cns.umass.edu:9283", cols = NULL)
}
\description{
wrapper function for char.retrieve
}
