% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{ls.objects}
\alias{ls.objects}
\title{Customization of ls function}
\usage{
ls.objects(pos = 1, pattern, alpha = F, head = FALSE, n = 10)
}
\description{
Customization of ls function, that includes size (memory) of all objects or data in current session environment.
}
\keyword{ls}

