% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge.sp}
\alias{merge.sp}
\title{internal, temporary replication of merge function to explicitly indicate merging of spatial object with a data.frame}
\usage{
\method{merge}{sp}(x, y, by = intersect(names(x), names(y)), by.x = by,
  by.y = by, all.x = TRUE, suffixes = c(".old", ""),
  incomparables = NULL, ...)
}
\description{
perhaps remove?
}

