% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot (for ggplot)}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{\code{ggplot objects}}

\item{plotlist}{\code{list of ggplot objections} (optional)}

\item{cols}{\code{numeric} number of columns, defaults to 1}
}
\description{
Function for combining multiple ggplots, without using facet. From R Cookbook for Graphics, http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2); and published at http://peterhaschke.com/Code/multiplot.R
}

