% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.model.list}
\alias{predict.model.list}
\title{predict from list of models}
\usage{
\method{predict}{model.list}(model.list, new.data, id.col = "site_no",
  year.col = "year", period.col = "season", bias.correction = NULL)
}
\description{
x
}

