% Generated by roxygen2 (4.0.1): do not edit by hand
\name{retransform}
\alias{retransform}
\title{retransform dependent variable}
\usage{
retransform(o, bias.correction = NULL, residuals = NULL,
  group = "site_no")
}
\description{
usually rever from log scale.  option for smearing bias correction
}

