% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/prediction.R
\name{stochastic.correction}
\alias{stochastic.correction}
\title{bias correction using stochastically generated residuals}
\usage{
stochastic.correction(y, residuals, retransform = exp,
  correction.method = "rnorm", correction.summary = "mean",
  interval = c(0.05, 0.5, 0.95), resid.mean = NA, correction.iter = 100,
  correction.seed = NULL)
}
\description{
resid are generated either by random generation from normal distribution, or bootstrapping observed residuals
}
