% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/assign_period.R
\name{to_month_date}
\alias{to_month_date}
\title{Associates calendar dates with a date representing its month.}
\usage{
to_month_date(d)
}
\arguments{
\item{d}{\code{vector of dates}}
}
\value{
\code{vector of dates}
}
\description{
Given a date, determines which month it falls into, and assigns the date corresponding to the first calendar day of the month.
}
\seealso{
\code{\link{conteStreamflow::to.water.year}}, \code{\link{conteStreamflow::to.season}}
}
