% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/assign_period.R
\name{to_wateryear_date}
\alias{to_wateryear_date}
\title{Associate calendar dates a date object or numeric year, representing its water year.}
\usage{
to_wateryear_date(d)
}
\arguments{
\item{d}{\code{vector of dates}}
}
\description{
Given a date, determines which water year it falls into, and assigns it a date corresponding to the last calendar day of the water year
}
\keyword{dates,}
\keyword{water}
\keyword{year}
