% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_weather_grid_livneh.R
\name{weather.grid.livneh.coords}
\alias{weather.grid.livneh.coords}
\title{determine coordinates (and regions) livneh weather grid(s) from ftp dir and file names}
\usage{

  weather.grid.livneh.coords(ftp.url = "ftp://gdo-dcp.ucllnl.org/pub/dcp/archive/OBS/livneh2014.1_16deg/ascii/daily/")
}
\description{
determine coordinates (and regions) livneh weather grid(s) from ftp dir and file names
}

