% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_weather_grid_livneh.R
\name{weather.grid.livneh.create}
\alias{weather.grid.livneh.create}
\title{turn points into polygons for livneh weather grid}
\usage{
weather.grid.livneh.create(grid.points, shapefile.dir = NULL)
}
\description{
turn points into polygons for livneh weather grid
}

