% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weather.retrieve}
\alias{weather.retrieve}
\title{import weather}
\usage{
weather.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  weather.pre.agg.function = (conteStreamflow::weather.pre.agg.function),
  weather.agg.function = (conteStreamflow::weather.agg.function),
  weather.lag.function = (conteStreamflow::weather.lag.function),
  template.date = NULL, template.period = NULL, cols.weather = NULL)
}
\description{
import weather
}

