% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weather_retrieve.R
\name{weather.retrieve2}
\alias{weather.retrieve2}
\title{import weather}
\usage{
weather.retrieve2(gages.spatial, set = "livneh_1950_2013",
  periods = c("monthly", "annual"),
  weather.pre.agg.function = (conteStreamflow::weather.pre.agg.function),
  weather.agg.function = (conteStreamflow::weather.agg.function),
  weather.lag.function = (conteStreamflow::weather.lag.function),
  template.period = NULL)
}
\description{
import weather
}

