% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates_weather.R
\name{weather.sets.retrieve}
\alias{weather.sets.retrieve}
\title{retrieve weather, based on specified set}
\usage{
weather.sets.retrieve(set = "mauer_1949_2010")
}
\description{
placeholder
}

