% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/weather_retrieve.R
\name{weather_antecedent}
\alias{weather_antecedent}
\title{calculate antecedent precip for set number of days}
\usage{
weather_antecedent(x, ante_days_precip_e = c(7, 30, 60),
  ante_days_pet = c(30, 60, 90), ante_days_gdd = c(30, 60, 90))
}
\description{
calculate antecedent precip for set number of days
}
