% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/weather_retrieve.R
\name{weather_fetch_ante}
\alias{weather_fetch_ante}
\title{get the antecedent precip leading up to high precip events}
\usage{
weather_fetch_ante(x_period, x_calc, j, date_col = "max_precip_e_date",
  ante_days_precip_e, ante_days_pet, ante_days_gdd)
}
\description{
get the antecedent precip leading up to high precip events
}
