% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/weather_retrieve.R
\name{weather_lag}
\alias{weather_lag}
\title{weather lag}
\usage{
weather_lag(y, lag1_cols = c("n_days_weather", "precip_e_total",
  "precip_raw_total", "pet_total", "gdd_total", "tmax_mean", "tmin_mean",
  "swe_day_total", "frozen_day_total"), lag2_cols = c("n_days_weather",
  "precip_e_total", "precip_raw_total", "pet_total", "tmax_mean", "tmin_mean",
  "swe_day_total"))
}
\description{
weather lag
}
