% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/weather_retrieve.R
\name{weather_retrieve}
\alias{weather_retrieve}
\title{import weather}
\usage{
weather_retrieve(sites, periods = c("day", "month", "wateryear"),
  alter_temp = NULL, alter_precip = NULL)
}
\description{
import weather
}
