% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/weather_retrieve.R
\name{weather_retrieve_alter}
\alias{weather_retrieve_alter}
\title{wrapper function to retrieve weather, alter it by prescribed amounts at the daily step, and aggregate}
\usage{
weather_retrieve_alter(sites, periods = c("day", "month", "wateryear"),
  alter_temp_vector = NULL, alter_precip_vector = NULL, save_dir = NULL,
  save_zero_change = F)
}
\description{
wrapper function to retrieve weather, alter it by prescribed amounts at the daily step, and aggregate
}
