% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utilities_cache.R
\name{cache.check.dir}
\alias{cache.check.dir}
\title{Checks that cache exist and contains all necessary subfolders}
\usage{
cache.check.dir(cache.dir)
}
\arguments{
\item{cache.dir}{\code{character}  character of cache directory location}
}
\value{
\code{boolean}
}
