% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clifs_model_functions.R
\name{clifs.model.list}
\alias{clifs.model.list}
\title{CliFSSS model create model list}
\usage{
clifs.model.list(df, period.type = "month", formula, sort = T,
  model.type = "lm")
}
\arguments{
\item{df}{\code{data.frame}}

\item{period.type}{\code{character}}

\item{formula}{\code{character}}
}
\value{
sort \code{boolean}
}
\description{
creates a list of models, one for each period
}

