% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/clifs_model_functions.R
\name{clifs.summary}
\alias{clifs.summary}
\title{prints overall summary of CliFSSS model list performance for each period}
\usage{
clifs.summary(ml)
}
\arguments{
\item{ml}{\code{list} model list, of  \code{lm} or \code{merMod} class models}
}
\description{
prints summary of CliFSSS model performance for each period
}
