% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/calibration_sets.R
\name{create.calib}
\alias{create.calib}
\title{create calibration and validation sets}
\usage{
create.calib(d.records, valid.type = "gages.years", valid.fraction = 0.5,
  seed = NULL, id.col = "site_no", year.col = "year",
  select.gages = NULL, select.years = NULL)
}
\description{
x
}
