% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/flow_retrieve.R
\name{create.cols.flow}
\alias{create.cols.flow}
\title{col names for flow stats}
\usage{

  create.cols.flow(flow.pre.agg.function = (conteStreamflow::flow.pre.agg.function),
  flow.agg.function = (conteStreamflow::flow.agg.function))
}
\description{
col names for flow stats
}
