% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weather_retrieve.R
\name{create.cols.weather}
\alias{create.cols.weather}
\title{col names for weather metrics}
\usage{

  create.cols.weather(weather.pre.agg.function = (conteStreamflow::weather.pre.agg.function),
  weather.agg.function = (conteStreamflow::weather.agg.function),
  weather.lag.function = (conteStreamflow::weather.lag.function))
}
\description{
col names for our processed/saved weather metrics
}

