% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.cols.weather}
\alias{create.cols.weather}
\title{col names for weather metrics}
\usage{

  create.cols.weather(agg.function.weather = (conteStreamflow::agg.function.weather))
}
\description{
col names for our processed/saved weather metrics
}

