% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flow.retrieve}
\alias{flow.retrieve}
\title{import flow}
\usage{
flow.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  flow.pre.agg.function = (conteStreamflow::flow.pre.agg.function),
  flow.agg.function = (conteStreamflow::flow.agg.function),
  template.date = NULL, template.period = NULL, return.spatial = F)
}
\description{
import flow from nwis web service, and aggregate to various metrics and by various periods
}

