% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/goodness.R
\name{goodness.boxplot}
\alias{goodness.boxplot}
\title{goodness of fit boxplots}
\usage{
goodness.boxplot(o.goodness, measure.vars = c("NSEff.real", "NSEff.log",
  "bias", "percent.bias"), period.name = "season")
}
\description{
x
}
