% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/char_retrieve.R
\name{impound.retrieve}
\alias{impound.retrieve}
\title{impound retrieve}
\usage{
impound.retrieve(gages.spatial,
  server.url = "http://felek.cns.umass.edu:9283", cols = NULL)
}
\description{
wrapper function for char.retrieve
}

