% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/utilities.R
\name{ls.objects}
\alias{ls.objects}
\title{Customization of ls function}
\usage{
ls.objects(pos = 1, pattern, alpha = F, head = NULL)
}
\description{
Customization of ls function, that includes size (memory) of all objects or data in current session environment.
}
\keyword{ls}
