% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.model.list}
\alias{predict.model.list}
\title{predict continuous timeseries from a set (list) of models by period}
\usage{
\method{predict}{model.list}(model.list, new.data, id.col = "site_no",
  period.type = "season", dep.var = "flow.mean", retransform = exp,
  smearing.correction = T, rnorm.correction = T, bootstrap.correction = T,
  correction.summary = "mean", correction.iter = 100,
  correction.seed = NULL)
}
\description{
x
}

