% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/create_weather_grid_livneh.R
\name{weather.grid.livneh.create}
\alias{weather.grid.livneh.create}
\title{turn points into polygons for livneh weather grid}
\usage{
weather.grid.livneh.create(grid.points, shapefile.dir = NULL)
}
