% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weather.grid.livneh.create}
\alias{weather.grid.livneh.create}
\title{gather regions and coordinates for livneh weather grid(s)}
\usage{
weather.grid.livneh.create(grid.coords,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs",
  shapefile.dir = NULL)
}
\description{
gather regions and coordinates for livneh weather grid(s)
}

