% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/create_weather_grid_livneh.R
\name{weather.grid.livneh.points}
\alias{weather.grid.livneh.points}
\title{turn coordinates into points for livneh weather grid(s)}
\usage{
weather.grid.livneh.points(grid.coords,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")
}
